llm_microservice/requirements.txt
fastapi==0.115.6
huggingface-hub==0.27.1
pydantic==2.10.4
python-dotenv==1.0.1
uvicorn==0.34.0